fix(conformance): keep the ByteRange gap offset as a long
CheckByteRange still narrowed c+d to an int for the under-coverage gap check. The cast was safe,
but only because the cdSum > fileLength branch returns before reaching it, which bounds the value
by an int — an implicit coupling between two guards several lines apart that would break silently
if either one moved. It is a long now, and the comparison against a revision's int XrefOffset
widens rather than truncating.

Also a correction to how the previous commit described the impact. The reader holds the whole
file in a ReadOnlyMemory<byte>, whose Length is an int, so it cannot open a genuine 2 GB-plus
document at all today — that is what issue #105's ">2GB" item covers. The truncation was
nonetheless reachable, because /ByteRange values are author-controlled and a small malformed or
hostile document can claim any offset it likes. With the old cast, a claimed offset of
4,294,967,296 wrapped to 0, so CheckByteRange's `c <= 0` guard returned early and the §6.4.3-1
coverage check was silently skipped. With the values intact, cdSum > fileLength now fires the
violation correctly. So the fix turns a check that quietly did nothing into one that reports —
which matters for malformed input now, and for large files whenever the reader supports them.

+++ src/VellumPdf.Conformance/Rules/Signature/SignatureRule.cs
@@ -196,7 +196,10 @@ internal sealed class SignatureRule : IConformanceRule
         if (cdSum == fileLength)
             return; // exact coverage — compliant
 
-        var gapStart = (int)cdSum;
+        // Stays long rather than narrowing. The cast that used to be here was safe only because the
+        // cdSum > fileLength branch above returns first, which bounds it by an int — a coupling
+        // that would break silently the moment either guard moved.
+        var gapStart = cdSum;
         foreach (var rev in context.Revisions)
         {
             if (rev.XrefOffset >= gapStart && rev.XrefOffset < fileLength)
